// src/utils/types.ts
export type ThemeMode = 'dark' | 'light';

export type RecordTest = {
  jam_ujian: string;
  jenis_ujian: string;
  kode_ruang: string;
  listening: string;
  nama_peserta: string;
  nama_ruang: string;
  no_id_peserta: string;
  no_tes: string;
  reading: string;
  score: string;
  structure: string;
  tgl_ujian: string;
};